requirements.txt
watchdog==2.1.9
Pillow